pyproject.toml
[project]
name = "antigravity-dev"
version = "0.1.0"
description = "A repo-aware, sandboxed, multi-agent AI development platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & Web
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.1.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    # Task Queue
    "dramatiq[redis]>=1.17.0",
    # Data Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # LLM Providers
    "openai>=1.57.0",
    "google-generativeai>=0.8.0",
    # Docker/Sandbox
    "docker>=7.1.0",
    # CLI
    "typer[all]>=0.14.0",
    "rich>=13.9.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "tenacity>=8.2.0",
    "cryptography>=42.0.0",
    "honcho>=2.0.0",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    # Schema Sync (Pydantic -> TypeScript)
    "datamodel-code-generator>=0.26.0",
    # Testing utilities
    "httpx>=0.28.0",
    "faker>=33.0.0",
    "factory-boy>=3.3.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
gravity = "backend.scripts.gravity_cli:app"
sync-schema = "backend.scripts.sync_schema:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["libs", "backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = [".", "libs", "backend"]

[tool.mypy]
python_version = "3.11"
strict = true
